+++ docs/api_doc.rst
@@ -5,7 +5,7 @@ API Documentation
 This is the detailed documentation of ``BaseMPFTPServer``, which
 holds all the functionality you gain by ``PyTest local FTP Server``.
 
-.. currentmodule:: pytest_localftpserver.plugin
+.. currentmodule:: pytest_localftpserver.servers
 
 .. autosummary::
    :nosignatures:
